code-quality: move validate args
